// src/net/generated/models.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: models.go

export interface Variables {
  id: number /* int32 */;
  varname: string;
  value: string;
  information: string;
  ts: string;
}
export interface CharacterBind {
  id: number /* uint32 */;
  slot: number /* int32 */;
  mapId: number /* uint32 */;
  x: number /* float64 */;
  y: number /* float64 */;
  z: number /* float64 */;
  heading: number /* float64 */;
}
export interface CharacterWallet {
  characterId: number /* uint32 */;
  pokedollars: number /* uint32 */;
}
export interface CharacterData {
  id: number /* uint32 */;
  accountId: number /* int32 */;
  name: string;
  lastName: string;
  title: string;
  suffix: string;
  mapId: number /* uint32 */;
  y: number /* float64 */;
  x: number /* float64 */;
  z: number /* float64 */;
  heading: number /* float64 */;
  gender: number /* uint8 */;
  factionId: number /* uint16 */;
  class: number /* uint8 */;
  birthday: number /* uint32 */;
  lastLogin: number /* uint32 */;
  timePlayed: number /* uint32 */;
  gm: number /* uint8 */;
  deletedAt?: string;
  options?: string;
}